Split PandeiaCalculation() out of pandeia_interface.py and into pandeia_calculation.py since the files are getting massive.
